This code contains a simple Student Portal app.
This code consists of a simple Signup and Login Screen for student and teacher. The sign in information for teacher is hardcoded in Constants file for 2 teachers while student has to signup first. The signup credentials are stored in Shared Preferences. The student logins using these credentials. The app verifies the signup details. Then 2 fragments are implemented using View Pager. One for Insert Courses and one of View Courses. The students adds a course in Insert Courses fragment and saves it. This information is stored in Database. In View courses fragment, a Refresh button is implemented which, when clicked, fetches courses data from database and displays it in recycler view. When any item of recycler view list is clicked, It shows the details of that course in a new activity implemented using Interface. Same is for teacher but it has Insert Students and View Students in Viewpager.